+++ src/render/scene_mgr.h
@@ -46,8 +46,8 @@ struct SceneManager
   VkBuffer GetMeshInfoBuffer()  const { return m_meshInfoBuf; }
   std::shared_ptr<vk_utils::ICopyEngine> GetCopyHelper() { return  m_pCopyHelper; }
 
-  uint32_t MeshesNum() const {return m_meshInfos.size();}
-  uint32_t InstancesNum() const {return m_instanceInfos.size();}
+  uint32_t MeshesNum() const {return (uint32_t)m_meshInfos.size();}
+  uint32_t InstancesNum() const {return (uint32_t)m_instanceInfos.size();}
 
   hydra_xml::Camera GetCamera(uint32_t camId) const;
   MeshInfo GetMeshInfo(uint32_t meshId) const {assert(meshId < m_meshInfos.size()); return m_meshInfos[meshId];}
